remove isduplicate fn()
The isDuplicate fn() was unable to properly check and regulate the behavior of the cart. More work will need to be done to figure out the exact reason for its inability to do the one purpose it was designed for. Yet, I am the creator. Perhaps the vault lies not with the code but with the mind that produced it. Perhaps I am the bug...